Moving fakewidget to the internal subdirectory.
