+++ container/draw_test.go
@@ -24,9 +24,9 @@ import (
 	"github.com/mum4k/termdash/internal/draw"
 	"github.com/mum4k/termdash/internal/draw/testdraw"
 	"github.com/mum4k/termdash/internal/faketerm"
+	"github.com/mum4k/termdash/internal/fakewidget"
 	"github.com/mum4k/termdash/internal/widgetapi"
 	"github.com/mum4k/termdash/linestyle"
-	"github.com/mum4k/termdash/widgets/fakewidget"
 )
 
 func TestDrawWidget(t *testing.T) {
